# [glyf] Optimize retrieval of hmtx/vmtx/gvar